+++ src/components/board/cloneBoard.tsx
@@ -4,6 +4,7 @@ import canMoveKnight from "../figures/knight/canMoveKnight";
 import canMovePawn from "../figures/pawn/canMovePawn";
 import canMoveQueen from "../figures/queen/canMoveQueen";
 import canMoveRook from "../figures/rook/canMoveRook";
+import { boardType } from "./types";
 
 export const canMoveDict: { [key: string]: Function } = {
     'pawn': canMovePawn,
@@ -14,14 +15,12 @@ export const canMoveDict: { [key: string]: Function } = {
     'knight': canMoveKnight
 };
 
-type boardTypeWithoutMoving = { [key: string]: { color: string, type: string, canMove: Function } };
 
-
-export const cloneBoard = (boardToCopy: boardTypeWithoutMoving) => {
-    const board: boardTypeWithoutMoving = {};
+export const cloneBoard = (boardToCopy: boardType) => {
+    const board: boardType = {};
 
     for (const position in boardToCopy) {
-      board[position] = { color: boardToCopy[position].color, type: boardToCopy[position].type, canMove: canMoveDict[boardToCopy[position].type].bind(null, boardToCopy[position].color, position, board) }
+      board[position] = { color: boardToCopy[position].color, type: boardToCopy[position].type, position }
     }
 
     return board;
