+++ src/components/board/kingUnderAttack.tsx
@@ -1,15 +1,19 @@
+import { canMove } from "../../redux/slice/chess";
+import { boardType } from "./types";
+
 type boardTypeWithoutMoving = { [key: string]: { color: string, type: string, canMove: Function } };
 
-export const kingUnderAttack = (board: boardTypeWithoutMoving, color: string) => {
-    for (const position in board) {
-      if (!board[position] || board[position].color === color) continue;
 
-      const posibles = board[position].canMove(false);
+export const kingUnderAttack = (board: boardType, color: string) => {
+  for (const position in board) {
+    if (!board[position] || board[position].color === color) continue;
 
-      for (const posiblePosition of posibles) {
-        if (board[posiblePosition] && board[posiblePosition].type === 'king' && board[posiblePosition].color !== board[position].color) return true;
-      }
+    const posibles = canMove(board, position);
+
+    for (const posiblePosition of posibles) {
+      if (board[posiblePosition] && board[posiblePosition].type === 'king' && board[posiblePosition].color !== board[position].color) return true;
     }
-
-    return false;
   }
+
+  return false;
+}
